Fix crouching going inactive even when head is colliding
On the test level, if you crouch under the tunnel behind the stone ramp and let go of the crouch input, your speed goes back to normal and you can even sprint.

This is because the crouch ability keeps the head low, but the controller sets the crouch active ability without considering head collision.

I have added a condition to the crouch.set_active() call in the character_controller_3d.gd file that keeps the crouch ability active if the head is still crouching down.